Supporting carousel posts that start with a video
Looks like Insomniac is crashing on every carousel which is starting with a video.

Images are placed to a `id/zoomable_view_container`, carousels which start with an image are placed to a `id/carousel_image`.
But carousels that start with a video are placed to a `id/carousel_video_container`.

Looks like Issue #239 is reporting that.

This fix is untested, please triplecheck. Testing is not thrilling me. lol